KTOR-5980 Replace iconv with Foundation on darwin platform
