+++ ktor-io/jvm/src/io/ktor/utils/io/charsets/CharsetJVM.kt
@@ -174,6 +174,33 @@ internal actual fun CharsetDecoder.decodeBuffer(
     return charactersCopied
 }
 
+internal actual fun CharsetEncoder.encodeToByteArrayImpl1(
+    input: CharSequence,
+    fromIndex: Int,
+    toIndex: Int
+): ByteArray {
+    var start = fromIndex
+    if (start >= toIndex) return EmptyByteArray
+    val single = ChunkBuffer.Pool.borrow()
+
+    try {
+        val rc = encodeImpl(input, start, toIndex, single)
+        start += rc
+        if (start == toIndex) {
+            val result = ByteArray(single.readRemaining)
+            single.readFully(result)
+            return result
+        }
+
+        return buildPacket {
+            appendSingleChunk(single.duplicate())
+            encodeToImpl(this, input, start, toIndex)
+        }.readBytes()
+    } finally {
+        single.release(ChunkBuffer.Pool)
+    }
+}
+
 // -----------------------
 
 public actual typealias CharsetDecoder = java.nio.charset.CharsetDecoder
